Move *Codec from cdap-app-fabric to cdap-proto
