+++ cdap-app-fabric/src/main/java/co/cask/cdap/internal/app/ApplicationSpecificationCodec.java
@@ -27,6 +27,7 @@ import co.cask.cdap.api.worker.WorkerSpecification;
 import co.cask.cdap.api.workflow.WorkflowSpecification;
 import co.cask.cdap.app.ApplicationSpecification;
 import co.cask.cdap.data.dataset.DatasetCreationSpec;
+import co.cask.cdap.proto.codec.AbstractSpecificationCodec;
 import com.google.gson.JsonDeserializationContext;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
@@ -38,7 +39,7 @@ import java.lang.reflect.Type;
 import java.util.Map;
 
 /**
- *
+ * TODO: Move to cdap-proto
  */
 final class ApplicationSpecificationCodec extends AbstractSpecificationCodec<ApplicationSpecification> {
