+++ cdap-proto/src/main/java/co/cask/cdap/proto/codec/FlowSpecificationCodec.java
@@ -1,0 +1,65 @@
+/*
+ * Copyright © 2014 Cask Data, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+package co.cask.cdap.proto.codec;
+
+import co.cask.cdap.api.flow.FlowSpecification;
+import co.cask.cdap.api.flow.FlowletConnection;
+import co.cask.cdap.api.flow.FlowletDefinition;
+import co.cask.cdap.internal.flow.DefaultFlowSpecification;
+import com.google.gson.JsonDeserializationContext;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonObject;
+import com.google.gson.JsonParseException;
+import com.google.gson.JsonPrimitive;
+import com.google.gson.JsonSerializationContext;
+
+import java.lang.reflect.Type;
+import java.util.List;
+import java.util.Map;
+
+/**
+ *
+ */
+public final class FlowSpecificationCodec extends AbstractSpecificationCodec<FlowSpecification> {
+
+  @Override
+  public JsonElement serialize(FlowSpecification src, Type typeOfSrc, JsonSerializationContext context) {
+    JsonObject jsonObj = new JsonObject();
+
+    jsonObj.add("className", new JsonPrimitive(src.getClassName()));
+    jsonObj.add("name", new JsonPrimitive(src.getName()));
+    jsonObj.add("description", new JsonPrimitive(src.getDescription()));
+    jsonObj.add("flowlets", serializeMap(src.getFlowlets(), context, FlowletDefinition.class));
+    jsonObj.add("connections", serializeList(src.getConnections(), context, FlowletConnection.class));
+
+    return jsonObj;
+  }
+
+  @Override
+  public FlowSpecification deserialize(JsonElement json, Type typeOfT,
+                                       JsonDeserializationContext context) throws JsonParseException {
+    JsonObject jsonObj = json.getAsJsonObject();
+
+    String className = jsonObj.get("className").getAsString();
+    String name = jsonObj.get("name").getAsString();
+    String description = jsonObj.get("description").getAsString();
+    Map<String, FlowletDefinition> flowlets = deserializeMap(jsonObj.get("flowlets"), context, FlowletDefinition.class);
+    List<FlowletConnection> connections = deserializeList(jsonObj.get("connections"), context, FlowletConnection.class);
+
+    return new DefaultFlowSpecification(className, name, description, flowlets, connections);
+  }
+}
