+++ cdap-proto/src/main/java/co/cask/cdap/proto/codec/FlowletSpecificationCodec.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package co.cask.cdap.internal.app;
+package co.cask.cdap.proto.codec;
 
 import co.cask.cdap.api.Resources;
 import co.cask.cdap.api.flow.flowlet.FailurePolicy;
@@ -34,7 +34,7 @@ import java.util.Set;
 /**
  *
  */
-final class FlowletSpecificationCodec extends AbstractSpecificationCodec<FlowletSpecification> {
+public final class FlowletSpecificationCodec extends AbstractSpecificationCodec<FlowletSpecification> {
 
   @Override
   public JsonElement serialize(FlowletSpecification src, Type typeOfSrc, JsonSerializationContext context) {
