+++ cdap-proto/src/main/java/co/cask/cdap/proto/codec/SparkSpecificationCodec.java
@@ -1,12 +1,12 @@
 /*
  * Copyright © 2014 Cask Data, Inc.
- *
+ *  
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
  * the License at
- *
+ *  
  * http://www.apache.org/licenses/LICENSE-2.0
- *
+ *  
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
  * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
@@ -14,12 +14,9 @@
  * the License.
  */
 
-package co.cask.cdap.internal.app;
+package co.cask.cdap.proto.codec;
 
-import co.cask.cdap.api.flow.FlowSpecification;
-import co.cask.cdap.api.flow.FlowletConnection;
-import co.cask.cdap.api.flow.FlowletDefinition;
-import co.cask.cdap.internal.flow.DefaultFlowSpecification;
+import co.cask.cdap.api.spark.SparkSpecification;
 import com.google.gson.JsonDeserializationContext;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
@@ -28,38 +25,37 @@ import com.google.gson.JsonPrimitive;
 import com.google.gson.JsonSerializationContext;
 
 import java.lang.reflect.Type;
-import java.util.List;
 import java.util.Map;
 
 /**
  *
  */
-final class FlowSpecificationCodec extends AbstractSpecificationCodec<FlowSpecification> {
+public final class SparkSpecificationCodec extends AbstractSpecificationCodec<SparkSpecification> {
 
   @Override
-  public JsonElement serialize(FlowSpecification src, Type typeOfSrc, JsonSerializationContext context) {
+  public JsonElement serialize(SparkSpecification src, Type typeOfSrc, JsonSerializationContext context) {
     JsonObject jsonObj = new JsonObject();
 
     jsonObj.add("className", new JsonPrimitive(src.getClassName()));
     jsonObj.add("name", new JsonPrimitive(src.getName()));
     jsonObj.add("description", new JsonPrimitive(src.getDescription()));
-    jsonObj.add("flowlets", serializeMap(src.getFlowlets(), context, FlowletDefinition.class));
-    jsonObj.add("connections", serializeList(src.getConnections(), context, FlowletConnection.class));
+    jsonObj.add("mainClassName", new JsonPrimitive(src.getMainClassName()));
+    jsonObj.add("properties", serializeMap(src.getProperties(), context, String.class));
 
     return jsonObj;
   }
 
   @Override
-  public FlowSpecification deserialize(JsonElement json, Type typeOfT,
-                                       JsonDeserializationContext context) throws JsonParseException {
+  public SparkSpecification deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context)
+    throws JsonParseException {
     JsonObject jsonObj = json.getAsJsonObject();
 
     String className = jsonObj.get("className").getAsString();
     String name = jsonObj.get("name").getAsString();
     String description = jsonObj.get("description").getAsString();
-    Map<String, FlowletDefinition> flowlets = deserializeMap(jsonObj.get("flowlets"), context, FlowletDefinition.class);
-    List<FlowletConnection> connections = deserializeList(jsonObj.get("connections"), context, FlowletConnection.class);
+    String mainClassName = jsonObj.get("mainClassName").getAsString();
+    Map<String, String> properties = deserializeMap(jsonObj.get("properties"), context, String.class);
 
-    return new DefaultFlowSpecification(className, name, description, flowlets, connections);
+    return new SparkSpecification(className, name, description, mainClassName, properties);
   }
 }
