+++ cdap-proto/src/main/java/co/cask/cdap/proto/codec/WorkflowActionSpecificationCodec.java
@@ -1,22 +1,22 @@
 /*
  * Copyright © 2014 Cask Data, Inc.
- *  
+ *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
  * the License at
- *  
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- *  
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
  * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
  * License for the specific language governing permissions and limitations under
  * the License.
  */
+package co.cask.cdap.proto.codec;
 
-package co.cask.cdap.internal.app;
-
-import co.cask.cdap.api.spark.SparkSpecification;
+import co.cask.cdap.api.workflow.WorkflowActionSpecification;
+import co.cask.cdap.internal.workflow.DefaultWorkflowActionSpecification;
 import com.google.gson.JsonDeserializationContext;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
@@ -30,32 +30,30 @@ import java.util.Map;
 /**
  *
  */
-final class SparkSpecificationCodec extends AbstractSpecificationCodec<SparkSpecification> {
+public final class WorkflowActionSpecificationCodec extends AbstractSpecificationCodec<WorkflowActionSpecification> {
 
   @Override
-  public JsonElement serialize(SparkSpecification src, Type typeOfSrc, JsonSerializationContext context) {
+  public JsonElement serialize(WorkflowActionSpecification src, Type typeOfSrc, JsonSerializationContext context) {
     JsonObject jsonObj = new JsonObject();
 
     jsonObj.add("className", new JsonPrimitive(src.getClassName()));
     jsonObj.add("name", new JsonPrimitive(src.getName()));
     jsonObj.add("description", new JsonPrimitive(src.getDescription()));
-    jsonObj.add("mainClassName", new JsonPrimitive(src.getMainClassName()));
     jsonObj.add("properties", serializeMap(src.getProperties(), context, String.class));
 
     return jsonObj;
   }
 
   @Override
-  public SparkSpecification deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context)
-    throws JsonParseException {
+  public WorkflowActionSpecification deserialize(JsonElement json, Type typeOfT,
+                                                 JsonDeserializationContext context) throws JsonParseException {
     JsonObject jsonObj = json.getAsJsonObject();
 
     String className = jsonObj.get("className").getAsString();
     String name = jsonObj.get("name").getAsString();
     String description = jsonObj.get("description").getAsString();
-    String mainClassName = jsonObj.get("mainClassName").getAsString();
     Map<String, String> properties = deserializeMap(jsonObj.get("properties"), context, String.class);
 
-    return new SparkSpecification(className, name, description, mainClassName, properties);
+    return new DefaultWorkflowActionSpecification(className, name, description, properties);
   }
 }
